Capturing error messages of compiling and cssembed

+++ lessc.py
@@ -32,10 +32,21 @@ class CompileLessOnSave(sublime_plugin.EventListener):
 
         css_path = opts['css_path']
         less_path = opts['less_path']
+
         args = [lessc, '-x', less_path + 'style.less', css_path + 'style.css']
-        subprocess.call(args)
+        p = subprocess.Popen(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+	output, errors = p.communicate()
+	if errors:
+            sublime.error_message(errors)
+            return
+
         args = ['java', '-jar', opts['cssembed_path'], css_path + 'style.css', '-o', css_path + 'style_base64.css']
-        subprocess.call(args)
+        p = subprocess.Popen(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+	output, errors = p.communicate()
+	if errors:
+            sublime.error_message(errors)
+            return
+
         os.remove(css_path + 'style.css')
         os.rename(css_path + 'style_base64.css', css_path + 'style.css')
-        view.set_status('less', 'File saved and LESS compiled forfolder ' + folder_name)
+        view.set_status('less', 'File saved and LESS compiled for folder ' + folder_name)
